refactor normalization for metrics calculation
